+++ types/hotel.go
@@ -7,7 +7,7 @@ type Hotel struct {
 	Name     string               `bson:"name" json:"name"`
 	Location string               `bson:"location" json:"location"`
 	Rooms    []primitive.ObjectID `bson:"rooms" json:"rooms"`
-	Rating   float64              `bson:"rating" json:"rating"`
+	Rating   int                  `bson:"rating" json:"rating"`
 }
 
 type Room struct {
